+++ lib/twi/mock/message.rb
@@ -1,6 +1,12 @@
 module Twi
-  class Mock::Message < Resource
+  class Mock::Message < Message
     # @return [String] unique identifier
     def id = @params[:id]
+
+    def create
+      @id = "SM#{rand}"
+      @status = 'delivered'
+      # todo rescue and then set @code
+    end
   end
 end